Simple and MockAuthenticationProvider need to be in separate modules so either one can be loaded as part of a LibraryAuthenticationProvider.
